accuracy - f1 score: without RAG: f1 score:  0.7876588021778584 accuracy score:  0.7876588021778584
with RAG:
f1 score:  0.8130671506352088
accuracy score:  0.8130671506352087
